Consolodate and fix tests
